+++ pyaurorax/search/conjunctions/classes/search.py
@@ -518,3 +518,18 @@ class ConjunctionSearch:
         """
         url = "%s/%s" % (self.__aurorax_obj.api_base_url, self.__aurorax_obj.search.api.URL_SUFFIX_CONJUNCTION_REQUEST.format(self.request_id))
         return requests_cancel(self.__aurorax_obj, url, wait, poll_interval, verbose)
+
+    def describe(self):
+        """
+        Describe the conjunction search as an "SQL-like" string.
+
+        Returns:
+            The "SQL-like" string describing the conjunction search object
+        """
+        # make request
+        url = "%s/%s" % (self.__aurorax_obj.api_base_url, self.__aurorax_obj.search.api.URL_SUFFIX_DESCRIBE_CONJUNCTION_QUERY)
+        req = AuroraXAPIRequest(self.__aurorax_obj, method="post", url=url, body=self.query)
+        res = req.execute()
+
+        # return
+        return res.data